Move 'processed' serializers
