+++ django-rgd-imagery/rgd_imagery/serializers/__init__.py
@@ -1,3 +1,10 @@
 from .base import ImageMetaSerializer, ImageSerializer, ImageSetSerializer
+from .processed import ConvertedImageSerializer, RegionImageSerializer
 
-__all__ = ['ImageSerializer', 'ImageMetaSerializer', 'ImageSetSerializer']
+__all__ = [
+    'ConvertedImageSerializer',
+    'ImageSerializer',
+    'ImageMetaSerializer',
+    'ImageSetSerializer',
+    'RegionImageSerializer',
+]
